+++ esper/defs/values/FormIdDef.cs
@@ -14,7 +14,7 @@ namespace esper.defs {
         public FormIdDef(DefinitionManager manager, JObject src)
             : base(manager, src) { }
 
-        public override dynamic ReadData(PluginFileSource source, UInt16? dataSize) {
+        public override dynamic ReadData(PluginFileSource source, UInt32? dataSize) {
             UInt32 data = source.reader.ReadUInt32();
             return FormId.FromSource(source.plugin, data);
         }
