+++ esper/defs/values/UInt8Def.cs
@@ -14,7 +14,7 @@ namespace esper.defs {
         public UInt8Def(DefinitionManager manager, JObject src)
             : base(manager, src) {}
 
-        public override dynamic ReadData(PluginFileSource source, UInt16? dataSize) {
+        public override dynamic ReadData(PluginFileSource source, UInt32? dataSize) {
             return source.reader.ReadByte();
         }
